Renaming EntityAction feature
