+++ TimberAPI/EntityAction/IEntityAction.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace TimberbornAPI.EntityInstantiatorSystem
+{
+    public interface IEntityAction
+    {
+        void ApplyToEntity(GameObject entity);
+    }
+}
