--- TimberAPI/EntityInstantiator/IEntityInstantiator.cs
@@ -1,9 +1,0 @@
-using UnityEngine;
-
-namespace TimberbornAPI.EntityInstantiatorSystem
-{
-    public interface IEntityInstantiator
-    {
-        void Instantiate(GameObject gameObject);
-    }
-}
